requirements.txt
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==310
pyzbar==0.1.9
six==1.17.0
tzdata==2025.2
